refactor: rename every javascript file to kebab case
