+++ www/layouts/admin/index.tsx
@@ -13,8 +13,8 @@ import { memo, ReactNode, useCallback, useMemo } from "react";
 import { ToastContainer } from "react-toastify";
 
 import Dropdown from "@/components/dropdown";
-import { adminDroppableArrowProps } from "@/components/droppableArrow";
-import { AdminDroppableIndicator } from "@/components/droppableIndicator";
+import { adminDroppableArrowProps } from "@/components/droppable-arrow";
+import { AdminDroppableIndicator } from "@/components/droppable-indicator";
 import { Head } from "@/components/head";
 import { Sprite } from "@/components/sprite";
 import { appConfig } from "@/config/app";
