+++ www/layouts/public/announcements-slider.tsx
