+++ www/pages/foo.tsx
